Make calendar iframes responsively sized widths
We managed to make it responsive to month vs agenda view, but the styling had the calendars either go offscreen or use too little of the screen. Moving the iframes to `Content.js` and setting `iframe {width: 100%}` resolved the issue as far as I can tell.

It just occurred to me as I write this that this may make it harder to directly change the calendar in the CMS, but we could make a workaround if need be.