Fix dropdown styling (mostly)
The "Events" and "Blog" dropdown are now seperate and for the most
part styled right, but "Events" is no longer lit up and underlined
with the redish coor because I need to review vanilla JS.

Basically, it needs to know if a slug from an event page is in the
current URL of the page. If so, the "Events" tab is styled to
look selected.

Issue #12