Create "About Page" template and content
The "About Page" right now is just the `DefaultPage`
template with an accordian tacked in. It seems like
for all of the content that the librarians want
on the page that it should be a series of pages
in a dropdown, but for the time being all of
this stuff will just be in accordians.