Fix missing descriptions in dropdowns
Field names were mismatched, so the content didn't show up. However, I noticed
that something as long as the internet policy will
bleed over the size of a dropdown and under dropdowns
that are below. Pushing this up as that is another issue, but not as critical as these
mismatched fields.